Рефакторин
Изменено наследование
Изменен аджакс админа
Доработан tinymce
Другие мелкие исправления

+++ system/admin/controllers/types/item.php
@@ -1,6 +1,6 @@
 <?php
 
-class Admin_Controllers_Types_Item extends Ajax{
+class Admin_Controllers_Types_Item extends Parents_AjaxUpload{
 	/**
 	 * @var Admin_Models_Types_Item
 	 */
@@ -32,15 +32,16 @@ class Admin_Controllers_Types_Item extends Ajax{
             '{tree_name}',
             '{tree_pid}',
             '{id_value}',
-            
+            '{f_value}'
         );
+        
         $replace = array(
             empty($id) ? 'add' : 'update',
             $tree['title'],
             $tree['name'],
             empty($id) ? intval($_POST['pid']) : $tree['pid'],
             $tree['id'],
-            
+            $entity->getF()
         );
 
         $typeModel = new Admin_Models_Type();
@@ -62,8 +63,8 @@ class Admin_Controllers_Types_Item extends Ajax{
             for($i = 0; $i < $cnt; $i++){
                 $cntJ = count($fields[$i]['variants']);
 
-                $str = 'get'.ucfirst($fields[$i]['name']).'()';
-                $arr = explode(',', $entity->$str);
+                $str = 'get'.ucfirst($fields[$i]['name']);
+                $arr = explode(',', $entity->$str());
                 for ($j = 0; $j < $cntJ; $j++){
                     $toReplace[] = '{'.$fields[$i]['name'].'_'.$j.'_value}';
 
@@ -95,9 +96,34 @@ class Admin_Controllers_Types_Item extends Ajax{
         $title = $this->isAjax() ? strip_tags($_POST['title']) : $data['title'];
         $name = $this->isAjax() ? strip_tags($_POST['name']) : $data['name'];
         $pid = $this->isAjax() ? intval($_POST['pid']) : $data['pid'];
+        $validator = new Libs_Validator(array(
+            'title' => 'Титулка',
+            'name' => 'Имя',
+            'pid' => 'Ид родителя'
+        ));
+        $data = array(
+            'title' => $title,
+            'name' => $name,
+            'pid' => $pid
+        );
+        $valid = array(
+            'title' => array('required' => true),
+            'name' => array('required' => true),
+            'pid' => array('required' => true)
+        );
+        if(!$validator->isValid($data, $valid)){
+            if ($this->isAjax()){
+                $json = array(
+                    'error' => true,
+                    'mess' => $validator->getErrors()
+                );
+                $this->putJSON($json);
+            }
+            return $validator->getErrors();
+        }
 
         $entity = new Entity_Item();
-        
+        $entity->setF($this->isAjax() ? strip_tags($_POST['item_f']) : $data['item_f']);
         $id = $this->itemModel->addItem($title, $name, $pid, $entity);
 
         if ($this->isAjax()){
@@ -123,6 +149,24 @@ class Admin_Controllers_Types_Item extends Ajax{
     */
     public function deleteItem($id = null){
         $id = $this->isAjax() ? intval($_POST['id']) : $id;
+        $validator = new Libs_Validator(array('id' => 'Ид'));
+        $data = array('id' => $id);
+        $valid = array('id' => array('required' => true));
+        if(!$validator->isValid($data, $valid)){
+            if ($this->isAjax()){
+                $json = array(
+                    'error' => true,
+                    'mess' => $validator->getErrors()
+                );
+                $this->putJSON($json);
+            }
+            return $validator->getErrors();
+        }
+        $tree = $this->itemModel->getItem($id);
+        $entity = new Entity_Item();
+        $entity->init($tree);
+
+        
 
         $result = $this->itemModel->deleteItem($id);
 
@@ -142,16 +186,45 @@ class Admin_Controllers_Types_Item extends Ajax{
         $tree->setId($this->isAjax() ? strip_tags($_POST['id']) : $data['id']);
         $tree->setTitle($this->isAjax() ? strip_tags($_POST['title']) : $data['tree_title']);
         $tree->setName($this->isAjax() ? strip_tags($_POST['name']) : $data['tree_name']);
+        $id = $tree->getId();
+        $title = $tree->getTitle();
+        $name = $tree->getName();
+        $validator = new Libs_Validator(array(
+            'title' => 'Титулка',
+            'name' => 'Имя',
+            'pid' => 'Ид родителя'
+        ));
+        $data = array(
+            'title' => $title,
+            'name' => $name,
+            'id' => $id
+        );
+        $valid = array(
+            'title' => array('required' => true),
+            'name' => array('required' => true),
+            'id' => array('required' => true)
+        );
+        if(!$validator->isValid($data, $valid)){
+            if ($this->isAjax()){
+                $json = array(
+                    'error' => true,
+                    'mess' => $validator->getErrors()
+                );
+                $this->putJSON($json);
+            }
+            return $validator->getErrors();
+        }
 
         $entity = new Entity_Item();
-        
-        $id = $this->itemModel->updateItem($tree, $entity);
+        $entity->setF($this->isAjax() ? strip_tags($_POST['item_f']) : $data['item_f']);
+        $result = $this->itemModel->updateItem($tree, $entity);
 
         if ($this->isAjax()){
             $json = array();
-            if($id){
+            if($result){
                 $json['error'] = false;
-                $json['mess'] = 'Добавлено';
+                $json['mess'] = 'Обновлено';
+                $json['clear'] = false;
                 $json['callback'] = 'function callback(){reloadMenu();}';
             } else{
                 $json['error'] = true;
@@ -161,6 +234,6 @@ class Admin_Controllers_Types_Item extends Ajax{
             $this->putJSON($json);
         }
 
-        return $id;
+        return $result;
     }
 }
